Indicate "unfinished work areas" with TODO comments.
Applicants working on this project may wish to use Visual Studio's Task List feature to keep track of the work being asked of them. This commit introduces TODO comments in areas where the applicant should modify the code, allowing Visual Studio to present these areas as Tasks for the applicant.